fix: synchronize versions, strengthen CI, and resolve build failures
- Updated VERSION file to 0.3.6 to match project manifests.
- Enhanced scripts/verify-version-sync.sh to validate VERSION and cli-npm/package.json against Cargo.toml.
- Added scripts/test-version-sync.sh with automated tests for version synchronization logic.
- Updated version-integrity CI workflow to trigger on all relevant file changes.
- Resolved WASM build failure by consolidating MAX_IMPORT_SIZE constant.
- Ensured tempfile availability in dev-dependencies for tests across all configurations.
- Refactored src/wasm.rs by moving methods to src/wasm_ext.rs to satisfy the 500-LOC limit.
- Fixed formatting in src/index/hnsw.rs.
