+++ src/wasm_ext.rs
@@ -5,11 +5,13 @@
 // Redundant clones are intentional for WASM ownership semantics
 
 #[cfg(target_arch = "wasm32")]
-use js_sys::{Array, Function, Uint8Array};
+use js_sys::{Array, Function, Float32Array, Uint8Array};
 #[cfg(target_arch = "wasm32")]
 use tokio::sync::broadcast::error::RecvError;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
+#[cfg(target_arch = "wasm32")]
+use wasm_bindgen::JsCast;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen_futures::spawn_local;
 
@@ -19,6 +21,107 @@ use crate::wasm::WasmFramework;
 #[cfg(target_arch = "wasm32")]
 #[wasm_bindgen]
 impl WasmFramework {
+    /// Get framework metrics snapshot
+    pub async fn metrics_snapshot(&self) -> Result<JsValue, JsValue> {
+        let metrics = self.framework.metrics_snapshot().await;
+        let obj = js_sys::Object::new();
+        js_sys::Reflect::set(
+            &obj,
+            &"concepts_injected_total".into(),
+            &(metrics.concepts_injected_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"associations_created_total".into(),
+            &(metrics.associations_created_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"probes_total".into(),
+            &(metrics.probes_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"avg_probe_latency_ms".into(),
+            &metrics.avg_probe_latency_ms.into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"cache_hits_total".into(),
+            &(metrics.cache_hits_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"cache_misses_total".into(),
+            &(metrics.cache_misses_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"cache_evictions_total".into(),
+            &(metrics.cache_evictions_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"reservoir_steps_total".into(),
+            &(metrics.reservoir_steps_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"avg_reservoir_step_latency_us".into(),
+            &metrics.avg_reservoir_step_latency_us.into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"reservoir_nodes_active".into(),
+            &(metrics.reservoir_nodes_active as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"persist_ops_total".into(),
+            &(metrics.persist_ops_total as f64).into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        js_sys::Reflect::set(
+            &obj,
+            &"avg_persist_latency_ms".into(),
+            &metrics.avg_persist_latency_ms.into(),
+        )
+        .map_err(|_| JsValue::from_str("failed to set JS property"))?;
+        Ok(obj.into())
+    }
+
+    /// Process a temporal sequence and return the resulting hypervector bytes.
+    #[wasm_bindgen(js_name = processSequence)]
+    pub async fn process_sequence(&self, sequence: Array) -> Result<Box<[u8]>, JsValue> {
+        self.framework
+            .validate_sequence_length(sequence.length() as usize)
+            .map_err(to_js_error)?;
+
+        let mut parsed_sequence = Vec::with_capacity(sequence.length() as usize);
+        for item in sequence.iter() {
+            let step = item
+                .dyn_into::<Float32Array>()
+                .map_err(|_| JsValue::from_str("processSequence expects Float32Array items"))?;
+            parsed_sequence.push(step.to_vec());
+        }
+
+        let output = self
+            .framework
+            .process_sequence(&parsed_sequence)
+            .await
+            .map_err(to_js_error)?;
+        Ok(output.to_bytes().into_boxed_slice())
+    }
     /// Get framework stats
     pub async fn stats(&self) -> Result<JsValue, JsValue> {
         let stats = self.framework.stats().await.map_err(to_js_error)?;
